Reroute to Blazium Github
There's a handful more places to reroute to godot discussion pages that we'll need to recreate or something. I'm not sure that editing the breadcrumbs or layout actually did anything.

This includes a first pass over the config file which should mostly correctly fix github url generation and some Titles.